# src/main/resources/openapi/exchange_currency_api.yaml
openapi: 3.0.3
info:
  title: Exchange Currency API
  description: |-
    ** Exchange Currency API **
  version: 1.0.0
servers:
  - url: /exchangeCurrency/swagger-ui/index.html
tags:
  - name: exchangeCurrency
    description: create account and exchange currencies
paths:
  /createAccount:
    post:
      summary: creates currency exchange account
      operationId: createAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccount'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountCreated'
        '400':
          description: Invalid request body supplied
        '422':
          description: Request body contains errors
  /exchangeCurrency: #/{accountNumber}/{currencyCode}:
    patch:
      summary: exchange one currency for another
      operationId: exchangeCurrency
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccount'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountCreated'
        '400':
          description: Invalid request body supplied
        '422':
          description: Request body contains errors
components:
  schemas:
    CreateAccount:
      type: object
      properties:
        firstName:
          type: string
          nullable: false
        lastName:
          type: string
          nullable: false
        initialBalance:
          nullable: false
          $ref: '#/components/schemas/AccountBalance'
    AccountCreated:
      type: object
      properties:
        accountNumber:
          type: string
    AccountBalance:
      type: object
      properties:
        currencyCode:
          nullable: false
          $ref: '#/components/schemas/CurrencyCode'
        balance:
          type: number
          format: double
          nullable: false
    CurrencyCode:
      type: string
      enum:
        - USD
        - PLN